Moving from IntWritable to LongWritable

+++ mapreduce-testing-mrjob/src/main/java/ru/develbureau/mrtesting/mapreduce/AccessLogParserMapper.java
@@ -2,7 +2,6 @@ package ru.develbureau.mrtesting.mapreduce;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -16,12 +15,12 @@ import ru.develbureau.mrtesting.util.Util;
  * Date: 23.03.13
  * Time: 16:01
  */
-public class AccessLogParserMapper extends Mapper<LongWritable, Text, LoggedRequest, IntWritable>{
+public class AccessLogParserMapper extends Mapper<LongWritable, Text, LoggedRequest, LongWritable>{
 
     private static final Log LOG = LogFactory.getLog(AccessLogParserMapper.class);
 
     private final ApacheLogParser parser = new ApacheLogParser();
-    private final IntWritable one = new IntWritable(1);
+    private final LongWritable one = new LongWritable(1);
 
     @Override
     public void map(LongWritable key, Text value, Context context){
