+++ mapreduce-testing-mrjob/src/main/java/ru/develbureau/mrtesting/mapreduce/AccessLogReducer.java
@@ -1,6 +1,6 @@
 package ru.develbureau.mrtesting.mapreduce;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.NullWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -13,14 +13,14 @@ import java.io.IOException;
  * Date: 23.03.13
  * Time: 16:59
  */
-public class AccessLogReducer extends Reducer<LoggedRequest, IntWritable, NullWritable, Text>{
+public class AccessLogReducer extends Reducer<LoggedRequest, LongWritable, NullWritable, Text>{
 
     private final Text outputText = new Text();
 
     @Override
-    public void reduce(LoggedRequest loggedRequest, Iterable<IntWritable> hits, Context context) throws IOException, InterruptedException {
+    public void reduce(LoggedRequest loggedRequest, Iterable<LongWritable> hits, Context context) throws IOException, InterruptedException {
         long totalCount = 0;
-        for(IntWritable counter : hits){
+        for(LongWritable counter : hits){
             totalCount+=counter.get();
         }
         context.write(NullWritable.get(), prettyPrint(loggedRequest, totalCount));
